Make async functions implementation manual
