+++ Gir.toml
@@ -579,6 +579,12 @@ status = "generate"
         [[object.function.parameter]]
         name = "argv"
         string_type = "os_string"
+    [[object.function]]
+    name = "communicate_utf8_async"
+    ignore = true
+    [[object.function]]
+    name = "communicate_utf8_async_future"
+    ignore = true
 
 [[object]]
 name = "Gio.SubprocessLauncher"
